backend/requirements-alt.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Session==0.5.0
Flask-Compress==1.20
Flask-Limiter==3.5.1
Flask-Talisman==1.1.0
python-socketio==5.8.0
gevent>=24.2.1
gevent-websocket>=0.10.1
gunicorn==21.2.0
python-dotenv==1.0.0
# Machine Learning dependencies (kept as-is; verify numpy 2 compatibility)
scikit-learn==1.7.2
pandas>=2.0.3
numpy>=2.0.0,<2.3.0
joblib==1.3.2
xgboost==2.0.3
lightgbm==4.3.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
pytest>=8.2,<9
beautifulsoup4==4.12.2
pytest-flask==1.3.0
selenium==4.15.2
webdriver-manager==4.0.1
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.4
Flask-Caching==2.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
# Security
bcrypt==4.0.1
PyJWT==2.8.0
pyotp==2.9.0
cryptography>=41.0.0
# Error monitoring
sentry-sdk[flask]==2.13.0
# Async / HTTP client updated for ollama compatibility
httpx>=0.27.0
httpcore>=0.17.3
aiofiles==23.2.1
